+++ src/react-app/src/App.css
@@ -21,6 +21,19 @@ hr {
   z-index: 1;
 }
 
+.navbar {
+  position: fixed !important;
+  top: 0;
+  width: 100%;
+  height: 140px;
+  z-index: 1000;
+}
+
+.main-content {
+  padding-top: 140px !important;
+}
+
+
 .container {
   max-width: 600px; /* Limits container width */
   width: 100%; /* Keeps responsiveness */
@@ -30,7 +43,7 @@ hr {
   flex-direction: column;
   align-items: center; /* Centers content horizontally */
   justify-content: flex-start; /* Keeps items at the top */
-  padding: 20px;
+  padding: 20px auto;
   margin: 20px auto; /* Centers container */
   border-radius: 10px;
   box-shadow: 0px 0px 10px rgba(255, 255, 255, 0.2);
